+++ run.py
@@ -1,0 +1,4 @@
+from flask_health import app  # imports from __init__.py file within the package
+
+if __name__ == '__main__':
+    app.run(debug=True)
